+++ resources/lang/ro/location-types.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'district' => 'raion',
+    'municipality' => 'municipiu',
+    'autonomous_unit' => 'unitate teritorială autonomă',
+    'territorial_unit' => 'unitate administrativ-teritorială',
+    'sector' => 'sector',
+    'town' => 'oraș',
+    'village' => 'sat',
+];