+++ resources/lang/ru/location-types.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'district' => 'район',
+    'municipality' => 'муниципий',
+    'autonomous_unit' => 'автономное территориальное образование',
+    'territorial_unit' => 'административно-территориальная единица',
+    'sector' => 'сектор',
+    'town' => 'город',
+    'village' => 'село',
+];