+++ resources/lang/uk/location-types.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'district' => 'район',
+    'municipality' => 'муніципій',
+    'autonomous_unit' => 'автономне територіальне утворення',
+    'territorial_unit' => 'адміністративно-територіальна одиниця',
+    'sector' => 'сектор',
+    'town' => 'місто',
+    'village' => 'село',
+];