+++ src/Facades/Cuatm.php
@@ -12,10 +12,11 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * @method static Location|null findByCode(string $code)
+ * @method static Location|null findByStatisticCode(string $code)
  * @method static Location|null findBySlug(string $slug)
  * @method static Collection<int, Location> ofType(LocationType $type)
  * @method static Collection<int, Location> roots()
- * @method static Collection<int, Location> raioane()
+ * @method static Collection<int, Location> districts()
  * @method static Collection<int, Location> municipalities()
  * @method static Collection<int, Location> childrenOf(Location|int $parent)
  * @method static Collection<int, Location> tree()